# swagger/swagger.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: HomeNet
  description: |
    listings of properties for sale
  license:
    name: MIT
servers:
- url: http://localhost:3000
paths:
  /listings:
    get:
      summary: Show all listings
      operationId: listings
      tags:
      - listings
      parameters:
      - name: Type
        in: query
        description: filter by query type
        required: false
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Array of listings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponsePosts"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Create one listings
      operationId: createListing
      tags:
      - listings
      requestBody:
        description: the listing to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Post'
      responses:
        '201':
          description: The created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponsePost"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
    Post:
      type: object
      properties:
        title:
          type: string
        body:
          type: string
        userId:
          type: integer
    ResponsePost:
      allOf:
      - $ref: "#/components/schemas/Post"
      - properties:
          id:
            type: integer
          _id:
            type: string
            format: uuid
    ResponsePosts:
      type: array
      items:
        $ref: "#/components/schemas/ResponsePost"
    Error:
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string